On branch master  Your branch is up to date with 'origin/master'.  Changes to be committed: 	modified:   Ledstrip/Makefile 	modified:   Ledstrip/main.cpp 	new file:   Makefile 	modified:   Makefile.due 	modified:   Makefile.link 	modified:   Makefile.native
Updated the main.cpp with a brand new 2 co-op player mode!
Brand new UI design in terminal! View modes and keep track of the wrong inputs and if you were too late.
Startup is now explained in the terminal.
Also this complete Library is now updated with new Makefiles and is compatible with the newest version of hwlib AND bmptk